+++ tchannel/thrift/rw.py
@@ -40,11 +40,6 @@ from .module import ThriftRequest
 def load(path, service=None, hostport=None, module_name=None):
     """Loads the Thrift file at the specified path.
 
-    .. note::
-
-        This functionality is experimental and subject to change. We expect to
-        mark it as stable in a future version.
-
     The file is compiled in-memory and a Python module containing the result
     is returned. It may be used with ``TChannel.thrift``. For example,
 
@@ -180,18 +175,19 @@ class TChannelThriftModule(types.ModuleType):
         """
 
         self.service = service
-        self.module = module
         self.hostport = hostport
 
-        for service_cls in self.module.services:
+        self._module = module
+
+        for service_cls in self._module.services:
             name = service_cls.service_spec.name
             setattr(self, name, Service(service_cls, self))
 
     def __getattr__(self, name):
-        return getattr(self.module, name)
+        return getattr(self._module, name)
 
     def __str__(self):
-        return 'TChannelThriftModule(%s, %s)' % (self.service, self.module)
+        return 'TChannelThriftModule(%s, %s)' % (self.service, self._module)
 
     __repr__ = __str__
 
@@ -228,21 +224,21 @@ class Function(object):
     """
 
     __slots__ = (
-        'spec', 'func', 'service', 'request_cls', 'response_cls'
+        'spec', 'service', '_func', '_request_cls', '_response_cls'
     )
 
     def __init__(self, func_spec, service):
         self.spec = func_spec
-        self.func = func_spec.surface
         self.service = service
 
-        self.request_cls = self.func.request
-        self.response_cls = self.func.response
+        self._func = func_spec.surface
+        self._request_cls = self._func.request
+        self._response_cls = self._func.response
 
     @property
     def endpoint(self):
         """Endpoint name for this function."""
-        return '%s::%s' % (self.service.name, self.func.name)
+        return '%s::%s' % (self.service.name, self._func.name)
 
     @property
     def oneway(self):
@@ -266,13 +262,13 @@ class Function(object):
             )
 
         module = self.service._module
-        call_args = self.request_cls(*args, **kwargs)
+        call_args = self._request_cls(*args, **kwargs)
 
         return ThriftRWRequest(
             module=module,
             service=module.service,
             endpoint=self.endpoint,
-            result_type=self.response_cls,
+            result_type=self._response_cls,
             call_args=call_args,
             hostport=module.hostport,
         )
@@ -311,8 +307,8 @@ def register(dispatcher, service, handler=None, method=None):
         dispatcher.register(
             function.endpoint,
             handler,
-            ThriftRWSerializer(service._module, function.request_cls),
-            ThriftRWSerializer(service._module, function.response_cls),
+            ThriftRWSerializer(service._module, function._request_cls),
+            ThriftRWSerializer(service._module, function._response_cls),
         )
         return handler
 
@@ -327,7 +323,7 @@ def build_handler(function, handler):
     # function. It accepts one parameter for each exception defined on the
     # method and another parameter 'success' for the result of the call. The
     # success kwarg is absent if the function doesn't return anything.
-    response_cls = function.response_cls
+    response_cls = function._response_cls
     response_spec = response_cls.type_spec
 
     @gen.coroutine