+++ tchannel/tornado/connection.py
@@ -28,6 +28,8 @@ import sys
 import tornado.gen
 import tornado.iostream
 
+import tornado.queues as queues
+
 from .. import errors
 from .. import frame
 from .. import glossary
@@ -44,13 +46,6 @@ from ..messages.error import ErrorMessage
 from ..messages.types import Types
 from .message_factory import MessageFactory
 from .util import chain
-
-try:
-    import tornado.queues as queues  # included in 4.2
-    QueueEmpty = queues.QueueEmpty
-except ImportError:
-    import toro as queues
-    from Queue import Empty as QueueEmpty
 
 log = logging.getLogger('tchannel')
 
@@ -143,7 +138,7 @@ class TornadoConnection(object):
             while True:
                 message = self._messages.get_nowait()
                 log.warn("Unconsumed message %s", message)
-        except QueueEmpty:
+        except queues.QueueEmpty:
             pass
 
     def await(self):
@@ -330,7 +325,8 @@ class TornadoConnection(object):
         return self.connection.write(body)
 
     def close(self):
-        self.connection.close()
+        if not self.connection.closed():
+            self.connection.close()
 
     @tornado.gen.coroutine
     def initiate_handshake(self, headers):