+++ tests/test_singleton.py
@@ -18,4 +18,39 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-__all__ = ['ttypes', 'constants', 'KeyValue']
+from __future__ import (
+    absolute_import, division, print_function, unicode_literals
+)
+
+import pytest
+
+from tchannel import TChannel as AsyncTChannel
+from tchannel.singleton import TChannel
+from tchannel.errors import SingletonNotPreparedError
+
+
+def test_get_instance_is_singleton():
+
+    TChannel.reset()
+    TChannel.prepare('my-app')
+
+    assert TChannel.get_instance() is TChannel.get_instance()
+
+
+def test_must_call_prepare_before_get_instance():
+
+    TChannel.reset()
+
+    with pytest.raises(SingletonNotPreparedError):
+        TChannel.get_instance()
+
+
+def test_get_instance_returns_configured_tchannel():
+
+    TChannel.reset()
+    TChannel.prepare('my-app')
+
+    tchannel = TChannel.get_instance()
+
+    assert isinstance(tchannel, AsyncTChannel)
+    assert tchannel.name == 'my-app'